% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{copy_assets}
\alias{copy_assets}
\title{Copy assets from package to current directory}
\usage{
copy_assets(output_path = "./assets/template", overwrite_ok = FALSE)
}
\arguments{
\item{output_path}{The path to the directory where the assets should be
copied.}

\item{overwrite_ok}{Should files be overwritten?}
}
\description{
This function copies assets from the package to the current working
directory. The assets are located in the "inst/assets" directory of the
package.
}
