% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_bookdown_yml}
\alias{create_bookdown_yml}
\title{Create a `_bookdown.yml` file}
\usage{
create_bookdown_yml(path)
}
\arguments{
\item{path}{[`character(1)`] The directory path where the `_bookdown.yml`
file will be created.}
}
\value{
[`NULL`] Invisibly returns `NULL` after creating the file.
}
\description{
This function generates a `_bookdown.yml` file in the specified directory
with metadata about the last build time.
}
